==> 1. clean_data.ipynb <==
from tox21_project import data_initialization

import warnings
warnings.filterwarnings("ignore")

filename = './data/tox21_10k_data_all.sdf'
di = data_initialization.dataInitialization(filename)

# 1. 目标列
target_columns = ['SR-HSE', 'NR-AR', 'SR-ARE', 'NR-Aromatase', 'NR-ER-LBD', 'NR-AhR', 
                  'SR-MMP', 'NR-ER', 'NR-PPAR-gamma', 'SR-p53', 'SR-ATAD5', 'NR-AR-LBD']
# 2. 清除FW中的括号
di.clean_num(['FW'])

# 3. 改变数据类型
di.change_types(target_columns, float)

# 4. 合并重复行
di.merge_duplicate_target_rows('DSSTox_CID', target_columns)

# 5. 保存数据
di.save('./data/data_dups_removed.csv')
