==> 2. preprocess.ipynb <==
import pandas as pd
import numpy as np
from tox21_project import data_preprocessing as dp
from tqdm import tqdm
from rdkit import Chem
from rdkit.Chem import AllChem
from rdkit import DataStructs

# import warnings
# warnings.filterwarnings("ignore")

### 生成Morgan Fingerprints

def generate_morgan_fingerprints(data:pd.core.frame.DataFrame, radius:int=2) -> str:
    print(f'generating morgan fingerprints...\nradius:{radius}')
    # 1. 生成分子指纹X
    smiles = data['SMILES']
    mf = dp.MorganFingerprints(smiles)
    X = mf.transform(radius)
    # 此处必须指定index，不然下面concat会有大麻烦
    X = pd.DataFrame(X, columns=['fp_' + str(i + 1) for i in range(0, X.shape[1])], index=data.index)
    # 2. 将分子指纹按列添加到DataFrame中
    df = pd.DataFrame(data[['DSSTox_CID', 'SMILES']])
    df = pd.concat([df, X], axis=1)
    # 3. 导出为csv
    filename = 'data/morgan_fingerprints_' + str(radius) + '.csv'
    df.to_csv(filename)
    return filename

### Murcko骨架

def generate_murckoscaffold(data:pd.core.frame.DataFrame) -> str:
    print('generating murcko scaffold...')
    # 1. 生成骨架
    smiles = data['SMILES']
    ms = dp.MurckoScaffold(smiles)
    X = ms.transform()
    X = pd.DataFrame(X, columns=['MurckoScaffold'], index=data.index) # 只有一列数据也需要加入[]
    # 2. 拼接
    df = data[['DSSTox_CID', 'SMILES']]
    df = pd.concat([df, X], axis=1)
    # 3. 导出
    filename = 'data/murcko_scaffold.csv'
    df.to_csv(filename)
    return filename

### 从Murcko骨架中生成分子指纹

def generate_morgan_fingerprints_from_murckoscaffold(radius:int=2) -> str:
    print('generating morgan fingerprints from murcko scaffold...')
    # 1. 生成指纹
    data = pd.read_csv('data/murcko_scaffold.csv', index_col=0)
    ms_smiles = data['MurckoScaffold'].values
    mf = dp.MorganFingerprints(ms_smiles)
    X = mf.transform(radius)
    X = pd.DataFrame(X, columns=['fp_' + str(i + 1) for i in range(0, X.shape[1])], index=data.index)
    # 2. 拼接
    df = data[['DSSTox_CID','SMILES','MurckoScaffold']]
    df = pd.concat([df, X], axis=1)
    # 3. 导出
    filename = 'data/morgan_fingerprints_from_murckoscaffold_' + str(radius) + '.csv'
    df.to_csv(filename)
    return filename

### 分子描述符

def generate_molecular_descriptors(data:pd.core.frame.DataFrame) -> str:
    print('generating molecular descriptors...')
    # 1. 取出数据
    smiles = data['SMILES']
    md = dp.MolecularDescriptors(smiles)
    de = md.get_descriptors()
    de = pd.DataFrame(de, columns=[str(name) for name, fun in md.descriptors], index=data.index)
    # 2. 合并
    df = data[['DSSTox_CID', 'SMILES']]
    df = pd.concat([df, de], axis=1)
    # 3. 导出
    filename = 'data/molecular_descriptors.csv'
    df.to_csv(filename)
    return filename

### 计算与已知毒性的相似度

def compute_similarity_scaffold_known_toxic(radius:int=2) -> str:
    print('compute similarity scaffold known toxic...')
    knownToxic = ['C1=CC=C(C=C1)N', 'C1=CC=C(C=C1)CBr', 'C1=CC=C(C=C1)CI', 'C1=CC=C(C=C1)CCl', 'C1=CC=C(C=C1)CF',
                 'CNN', 'C1=CC=C(C=C1)[N+](=O)[O-]', 'NN=O', 'C1=CC=C2C(=C1)C=CC3=CC=CC=C32', 'C1=CC(=O)C=CC1=O',
                 'C(CCl)SCCCl', 'C(=S)(N)N']
    # 1. 生成已知毒物的分子指纹
    knownToxic_mf = dp.MorganFingerprints(knownToxic)
    knownToxic_mf.transform(radius)
    # 2. 生成分子骨架的指纹
    scaffold = pd.read_csv('data/murcko_scaffold.csv', index_col=0)
    scaffold_mf = dp.MorganFingerprints(scaffold['MurckoScaffold'])
    scaffold_mf.transform(radius)
    # 3. 计算两个指纹的相似度
    res = [[DataStructs.FingerprintSimilarity(tox, scaf) for tox in knownToxic_mf.fingerprints] for scaf in tqdm(scaffold_mf.fingerprints)]
    df = pd.DataFrame(res, columns=knownToxic, index=scaffold.index)
    # 4. 导出
    filename = 'data/similarity_scaffold_known_toxic.csv'
    df.to_csv(filename)
    return filename

### 清理每个文件中全为0或其中包含单个NULL的列

def clean_data_zeros_nan(filename) -> None:
    print(f'clean data zeros and nan files...\nfile:{filename}')
    df = pd.read_csv(filename, index_col=0)
    # 1. 去除全为0的列
    res = df.apply(lambda x: (x == 0).all()).to_dict() # 返回column:bool
    for k, v in res.items():
        if v:
            del df[k]
    # 2. 去除null列
    res = df.apply(lambda x : (x.isnull()).any()).to_dict()
    for k, v in res.items():
        if v:
            del df[k]
    df.to_csv(filename)

### 主程序

data = pd.read_csv('data/data_dups_removed.csv', index_col=0)
radius = 10
all_files = []

# Morgan Fingerprints
all_files.append(generate_morgan_fingerprints(data, radius))

# Murcko骨架
all_files.append(generate_murckoscaffold(data))

# 从Murcko骨架中生成分子指纹
all_files.append(generate_morgan_fingerprints_from_murckoscaffold(radius))

# 分子描述符
all_files.append(generate_molecular_descriptors(data))

# 计算与已知毒性的相似度
all_files.append(compute_similarity_scaffold_known_toxic(radius))

# 清理文件数据
for file in all_files:
    clean_data_zeros_nan(file)
